pyproject.toml
[project]
name = "kss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=25.1.0",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.11.0.post1",
    "huggingface-hub>=0.34.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "openai>=1.97.1",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "transformers>=4.54.0",
    "yfinance>=0.2.65",
]
